Add support for Windows shell escaping
Refactor the code base to make it easier to support escaping on
different operating systems, and add support for escaping arguments on
Windows. The existing test suites have been updated to ensure all this
works correctly.

To ensure that the different parts of the system work together, all
dependent modules are stubbed using SinonJS [1], which has been added as
a devDependency.

--
1. https://sinonjs.org