Improve docs CI job time
Add a Gemfile and Gemfile.lock to the repository in the docs/ folder
matching the dependencies required for the documentation website.

This lockfile is now used in the test-and-lint.yml docs job to cache
and restore te cache. This cuts down the time this job takes from over 3
minutes to less than 20 seconds.

The .gitignore files have been updated to more accurately ignore Ruby/
Bundle/Jekyll directories. The .gitattributes file has been updated to
ignore Gemfile.lock in the diff. The .prettierignore file has been added,
primarily, to avoid formatting Ruby/Bundle/Jekyll files, but now also
specifies other generated files and folders to be ignored.