Improve error message for non-stringable values
Improve the existing error message when an argument cannot be converted
into a string (*) to make it clearer what went wrong and how this can be
solved. This is achieved by checking if the value is 1) a string or 2)
can be converted into a string using a `.toString` function.

Unrelatedly, this changes `assert(false, msg)` calls into
`assert.fail(msg)`.

--
*: "Cannot read property 'toString' of undefined"