Fuzz index.js against the printf command
Update the fuzz logic in index.fuzz.js to (additionally) test
`shescape.quote` against a real shell. The output of `shescape.quote` is
fed into the `printf` command [1], which should just output what it was
given (`prepareArgForPrintf` is used to alter the fuzz value a bit to
make this true, e.g. by escaping "%"). In practice, `printf` may change
some things, which `getExpctedPrintfString` converts the input into.

I choose `printf` because it is a relatively simple command to just get
back what you put in, so that if the fuzzer found an input that executes
arbirary commands, the output is different than expected. I didn't use
`echo` as it was a little bit inconsistent with handling whitespace.

The current state of the fuzzer may not be perfect w.r.t. the behaviour
of `printf`, but running it for a while I didn't encounter any problems.
Any improvements to the fuzzing logic w.r.t. `printf` are welcome, as
are potential fixes to avoid changes to the input made in
`prepareArgForPrintf`.

--
1. https://en.wikipedia.org/wiki/Printf_(Unix)