Add separate test code license
Add a separate license for test (and fuzz) code, namely the Unlicense
license [1]. This essentially means that anyone can take the test (and
fuzz)  code and do whatever they want with it.

Additionally, the fuzz logic is improved a  bit w.r.t. the behaviour of
printf, and some minor improvements to the Fuzzing section in the
Contributing Guidelines.

--
1. https://choosealicense.com/licenses/unlicense/