refactor: convert source from CommonJS to ECMAScript modules
This refactors the source code of this project from CommonJS to
ECMAScript modules, details on what this means for the source can be
found below. The distribution approach in package.json is based on
approach #2 of [1].

That is, package.json now specifies that this project is writtin in ESM
(`"type": "module"`) and how it can be used by both ESM and CommonJS
projects ("main" and "exports" fields). To provide support for CommonJS,
NCC [2] is used to transiple the ESM source back to CommonJS at publish
time. This file is ignored by git, but will be published to NPM.

Since v14 is the first major LTS version of Node with full ESM support,
the Contributing Guidelines have been updated to tell contributors to
use v14.16 or higher. Additionally, the Contributing Guidelines sees
some minor changes as well as updates to the fuzzing section.

## Source

The project source files now use `import`/`export` as opposed to
`require`/`module.exports`.

## Tests & Fuzzing

The test suites have been updated in two ways. First, as with the source
`require`/`module.exports` has been replaced by `import`/`export`.
Second, mocking of internal modules has been replaced by comparing
resulting values to the output of functions that are expected to be
called. This is slightly undesirable as an Error from a previously
mocked feature will now cause some tests to fail, however mocking
ES modules is currently not supported by Sinon.

As NYC does not currently support ECMAScript modules, I decided to
switch to c8 instead [3].

As for fuzzing, the fuzz source is unchanged as jsfuzz does not seem to
work with ECMAScript modules. This does mean that fuzz files must now
use the `.cjs` extension to tell Node that this is a CommonJS module.
This in turn requires that the source is transpiled before fuzzing,
hence the "prefuzz" npm script.

## Continuous integration

To be able to work nicely with ECMAScript modules the Continuous
Integration configuration has been updated to use Node v14.

--
1. https://nodejs.org/api/packages.html
   #packages_dual_commonjs_es_module_packages
2. https://www.npmjs.com/package/@vercel/ncc
3. https://github.com/bcoe/c8