chore: add StrykerJS
Add support for running StrykerJS [1] mutation tests on the Shescape
source. This includes adding it as a dependency, adding a script for
running it, updating the clean command to remove any artefacts generated
by it, and adding those artefacts to the .gitignore file.

Additionally, the coverage report is now outputted to /reports/coverage,
in-line with the /reports/mutation used by StrykerJS.

The Contributing Guidelines have also been updated to add a subsection
on running mutation tests. Additionally the structure has been slightly
altered, making "Testing" a level 2 section and both  "Fuzzing" and
"Mutation testing" a level 3 section.

Lastly, this re-adds /.nyc_output/ to the .gitignore and clean command
as it will still be generated by jsfuzz.

--
1. https://stryker-mutator.io/