refactor: simplify `escapeShellArgByPlatform`
Omit the check if a povided argument is a string, and instead always
check if the argument `isStringable` AND run `arg.toString()`. This is
safe as 1) `isStringable` checks if that is possible and 2) strings
are themselves stringable.

This possibility was found as a result of mutation testing, which showed
that

  if (typeof arg !== "string") {

can be mutation without tests noticing. In fact, this is very hard to
test as the branch is not necessary.