Upgrade package-lock.json to v2
Use npm version 7.7.6 to update package-lock.json to v2.  Updated the
Contributing Guidelines to reflect the fact that npm v7 is now required
for contributing to this project.

Also added an `.nvmrc` file pointing to node v15.14.0 to easily switch
to a compatible Node and npm version when using Node Version Manager.
And also updated the GitHub Actions Workflows to use Node v15.14 so it
uses npm v7.