Use local echo helper for fuzzing
Update the fuzzer to use a project-local echo-like helper, written in
JavaScript for Node.js, as an oracle to verify that shescape escapes
arguments properly. The echo utility simply reads the first argument it
receives when called like

  node path/to/echo.js "first argument"

and writes it back to std out.

This change would have a couple of advantages over the previous approach
1. It removes the `printf` prerequisite, which by extension ...
2. Enables fuzzing on Windows (untested as of creating this commit), and
3. Avoids having to account for `printf` behaviour in the fuzz target.