Fix fuzzing crash on Windows
Command arguments for Node.js on Windows are interpreted with
backslashes escaping double quotes (as if it's a JavaScript string). As
a result, arguments echod by the `test/fuzz/echo.js` script may have
characters missing. The fuzz target is updated to deal with this by
preparing the argument to be escaped.