Support configuring shell to fuzz
Update the fuzz logic to be able to configure the shell to fuzz through
an environment variable. To support users to easily set an environment
variable, dotenv [1] was added as a dependency to be able to configure
the environment variable through a `.env` file.

The Contributing Guidelines and `.gitignore` file have been updated
accordingly.

--
1. https://www.npmjs.com/package/dotenv