Fix fuzzing crash on Windows
... due to how PowerShell arguments are interpreted by NodeJS when there
is at least one space in the argument. This is the same problem as was
fixed in commit [1] for CMD.

The crash is included in the fuzz corpus to prevent regression in the
fuzzer as it took ~2 hours for the fuzzer to find this problem.

Additionally, update .gitattributes to hide diffs for the fuzz corpus.

--
1. 9011f7e8b5e3d744937f1179239cf57d330f4d80