Fix fuzzing crash on Windows
... due to how PowerShell arguments are interpreted by NodeJS when there
is no whitespace in the argument. This is a similar problem as was fixed
in commit [1] for arguments with a space in PowerShell.

The crash is included in the fuzz corpus to prevent regression in the
fuzzer as it took ~3 hours for the fuzzer to find this problem, a
related crash is also included for the special case where the slash is
at the end of the argument and would escape the final quote.

--
1. 445eb97706d51530e5d6e47aea07e60e92d9068d