Add property tests for unix.js & win.js
Add an initial property test suite for unix.js & win.js based on
existing unit tests that translate well to property tests but
restirected to only tests that make sense as property tests. E.g.
testing _how_ a certain character is escaped is not a "general" property
of these functions because how it will be escaped depends on the shell.