Remove Sinon
Remove the Sinon dependency and update the test suite using it to no
longer rely on it. The use of Sinon was limited to stubbing a call to a
standard library but the test is equally valid if the test just uses the
real value of the call to the standard library.