Add more property tests (#100)
* Add property tests for main.js

Add an initial property test suite for main.js based on existing unit
tests that translate well to property tests. The scenario of booleans,
undefineds, and objects lacking `.toString` are omitted because the unit
test suites are already exhaustive.

* Add property tests for unix.js & win.js

Add an initial property test suite for unix.js & win.js based on
existing unit tests that translate well to property tests but
restirected to only tests that make sense as property tests. E.g.
testing _how_ a certain character is escaped is not a "general" property
of these functions because how it will be escaped depends on the shell.

* Normalize property tests

- The number of runs.
- Interrupt time limit (just below the standard timeout of Mocha) - the
  custom time limit when running property tests is removed as well.
- Don't error if time limit is reached.
- Fileoverview of property test suites.

* Simplify, fix, and improve index.prop.js

- Organizing the tests using `describe`.
- Removing tests for invalid property tests (e.g. the output is not
  necessarily always longer than or equal to the input).
- Replacing the system-dependent `before` hook by a system-independent
  regular expression.
- Include options object when property testing the package API