Improve existing tests (#101)
* Remove Sinon

Remove the Sinon dependency and update the test suite using it to no
longer rely on it. The use of Sinon was limited to stubbing a call to a
standard library but the test is equally valid if the test just uses the
real value of the call to the standard library.

* Organize index.test.js

* Fix typo in constants.test.js

* Fix call signature of `unix.getDefaultShell` in main.test.js

* Reduce duplication of constants in unit & prop tests

* Simplify & organize main.test.js

* Improve index.prop.js

Increase accuracy of quotes pattern in quotes tests. Specifically the
test now ensures that the quotes match.

* Simplify and improve index.test.cjs

- Fix file overview description
- Simplify test assertions
- Update individual test descriptions
- Improve regular expression used for asserting quote functions