Combine old fuzz workflow with regular push checks
Rework the workflows for this project with the goal of merging the old
fuzz workflow into the "normal" checks to avoid having the fuzz job
duplicated between two workflows.

This is achieved by splitting the (now-removed) `push-checks.yml`
workflow into one for testing, one for linting, and one for things
related to the documentation. The lint and docs workflow simply contain
the lint and docs jobs previously found in `push-checks.yml`. The tests
workflow most closely resembles `push-checks.yml`, but with the lint and
docs jobs removed. The fuzz workflow is removed, the fuzz job is already
in the test workflow, but the test workflow (in its entirety) now runs
on a schedule as well, preserving the continuous fuzzing aspect of the
(now-removed) fuzz workflow. This means all tests are now run on a
schedule.