Fix fuzz error PowerShell (#112)
* Add PowerShell fuzz error to corpus

* Fix PowerShell fuzz error related to U+0085

Fix an error in the fuzzer code for PowerShell. The error relates to the
NEL character (U+0085) [1]. The error is the result of the fact that, in
PowerShell, the last character before the closing quote (`"`) is handled
differently if there's whitespace in the argument (see comments in the
`index.fuzz.cjs` file for more details). Since NEL is a whitespace
character this behaviour kicked in, but since NEL *isn't* a whitespace
character (`/\s/`) according to Node's RegExp engine an argument with
NEL wouldn't be prepared correctly by the fuzzer.

--
1. https://www.compart.com/en/unicode/U+0085