Add Docker instructions to the Contributing Guidelines
Add basic instructions for how a novice Docker user can use Docker when
contributing to Shescape.

Update the `.nvmrc` file to omit the "v" (which is compatible with the
spec) to allow for using it in the command that creates a Docker
container, in particular to specify the node version of the container.