Focus lint workflow
Update the lint.yml workflow to only run on pushes and pull requests
where files change that are linted/formatted (per the `lint` and
`format` scripts in package.json this is cjs, js, md, and yml files)
**or** when dependencies change (in case the changed dependency
influences the linting process - e.g. the current linter, prettier).

The lint workflow does not need to be run on a schedule as linting rules
are static (given the same dependencies), so if the linter passed in the
past it will always pass in the future.