Focus CI workflows (#122)
* Focus analysis workflow

Update the analysis.yml workflow file to only run on pushes and pull
requests where code is changed, or the configuration file for the
analysis is changed.

The analysis is stil run on a schedule in case the analysis tool has
more or improved rules that it scans for.

* Focus test workflow

Update the test.yml workflow to only run on pushes and pull request
where code (**.js), the node version for dev (.nvmrc), dependencies
(package-lock.json), transpile-related configuration (rollup.config.js)
test-related configuration (stryker.config.json) or the fuzz corpus
(test/fuzz/corpus/*) change.

Tests are still run on a schedule to give randomized test (fuzzing and
property tests) a chance to find bugs on a regular schedule.

* Focus lint workflow

Update the lint.yml workflow to only run on pushes and pull requests
where files change that are linted/formatted (per the `lint` and
`format` scripts in package.json this is cjs, js, md, and yml files)
**or** when dependencies change (in case the changed dependency
influences the linting process - e.g. the current linter, prettier).

The lint workflow does not need to be run on a schedule as linting rules
are static (given the same dependencies), so if the linter passed in the
past it will always pass in the future.