Setup benchmarking with Benchmark.js (#125)
* Add Benchmark.js as a dependency

Add a dependency to benchmark Shescape to the project,
Benchmark.js [1, 2]. This is the benchmark library for JavaScript used
by _the_ GitHub Action for Continuous Benchmarking [3].

* Add initial benchmark and a benchmarking command

Add an initial set of benchmarks for the `escapeShellArg` functions
(unix and win). Also add an npm command for running the benchmarks:

    npm run benchmark

Which will invoke the test/bench.js file, which currently contains all
benchmark-related code.

* Set up continous benchmarking

Set up continuous benchmarking using github-action-benchmark [3]
(unofficial), which will take the benchmark results of Benchmark.js [1]
and, using a cache, track the performance over time. The action is
configured to fail and create an alert (in the form of a comment) when a
(significant) performance degradation is detected.

--
1. https://www.npmjs.com/package/benchmark
2. https://benchmarkjs.com/
3. https://github.com/marketplace/actions/continuous-benchmark