Update CodeQL configuration (#134)
* Don't filter by paths `on: push` in code analysis workflow

Following this warning from CodeQL:

    **CodeQL**
    1 issue was detected with this workflow: Using on.push.paths can
    prevent Code Scanning annotating new alerts in your pull requests.

Outputted by `github/codeql-action/init@v1.0.24`.

* Include `./index.js` in the CodeQL analysis

Currently only the files inside the `src/` directory are being analyzed
due to the way CodeQL is configured. Update the configuration to include
the JavaScript file in the root dir of the project.

* Run CodeQL on all branches, not only `main` and PRs into `main`

Not really any reason to restrict this. The more code is analysed the
sooner problems will be uncovered.