Add script to update the changelog
Add a simple script that can be invoked to update the changelog. The
script alters the changelog *only* if there are unreleased changes, in
which case it sets the unreleased changes as changes released in the
version found (at time of running) in the package manifest (provided no
such release exists in the CHANGELOG).

Updated the Release Guidelines to recommend using this script.