Add script to update the index.js JSDoc
Add a simple script that can be invoked to update the JSDoc in index.js.
The script simply reads the version in the package manifest and writes
it to the JSDoc of index.js. It does not check whether this changes
anything or not.

Updated the Release Guidelines to recommend using this script.