Check if a merge is in progress before stashing
Update the pre-commit and pre-push hook to check if a merge is in
progress before stashing changes. This is needed because stashing during
a merge typically results in an unwanted repository state.

To support this, the `common.sh` script records whether there is a merge
in progress by checking if the `.git/MERGE_HEAD` file exists [1]. If the
file exists, the `IS_MERGING` variable is set to a non-empty string -
which is evaluated to true by the default (`-n`) testing type [2] - else
the `IS_MERGING` variable is set to an empty string - which is evaluated
to false by the default (`-n`) testing type [2].

--
1. https://stackoverflow.com/a/29101748
2. https://unix.stackexchange.com/a/447244