Add job to Automated Release workflow to commit a version bump
Add a new job to the Automated Releases workflow to automate the version
bump commit. To prevent having to push to the `main` branch (allowing
for full branch protections) and to encourage review before accepting an
automated change, the version bump is committed to a separate branch and
a Pull Request is opened through peter-evans/create-pull-request [1]. It
uses the new scripts added in the Release Guidelines to make the changes
for the version bump commit.

The existing jobs in the workflow are updated to only run on `push`
events, not changing their behaviour otherwise. The new job is only run
on `workflow_dispatch`, preventing it from running when a release is
being published.

I have tested the chagnes to the workflow in a private repository.

--
1. https://github.com/peter-evans/create-pull-request/tree/main