Improve git hooks (#128)
* Prevent restoring stashes if none were created

Add a utility script, `common.sh` for the git hooks with a function to
check if a stash was created during the hook execution. This is needed
because sometimes a stash isn't created (e.g. when it's not necessary)
in which case popping the stash could fail (previously prevented with
`|| true`) or unintentionally pop another stash.

To do this `common.sh` uses a helper function to get the stash count,
`get_stash_count` that uses a git command to get the number of stashes.
It redirects errors to `/dev/null` to hide unwanted output in case there
are 0 stashes. It either returns the git command output, or explicitly
returns "0" in case the output is empty (which means there was an error
because there were no stashes).

`git_stash_count` is invoked once on load to get the initial number of
stashes. A function is then exposed to check, dynamically, if a stash
has been created since the file was loaded. This function is used in
both the pre-push and pre-commit hook to check if a stash should be
popped after the checks have run.

* Use full names for flags/options to improve readability

* Stash unstaged *and* staged changes before pushing

Stashing the staged changes in addition to the unstaged changes before
pushing results in a repository state that 100% matches what is being
pushed.

* Check if a merge is in progress before stashing

Update the pre-commit and pre-push hook to check if a merge is in
progress before stashing changes. This is needed because stashing during
a merge typically results in an undesirable repository state.

To support this, the `common.sh` script exposes a function to check if a
merge is in progress by checking if the `.git/MERGE_HEAD` file exists
[1]. If the file exists, a non-empty string - which is evaluated to true
by the default (`-n`) testing type [2] - is returned, else the return
value is set to an empty string - which is evaluated to false by the
default (`-n`) testing type [2].

* Maintain index (staged/unstaged) after popping stash

Update the `git stash pop` flags for the pre-push hook to maintain the
index (i.e. whether a file was staged or not) of the stashed files. It's
not used in the pre-commit hook because there won't be any staged files
after committing.

--
1. https://stackoverflow.com/a/29101748
2. https://unix.stackexchange.com/a/447244