Further automate release process (#137)
* Add script to update the changelog

Add a simple script that can be invoked to update the changelog. The
script alters the changelog *only* if there are unreleased changes, in
which case it sets the unreleased changes as changes released in the
version found (at time of running) in the package manifest (provided no
such release exists in the CHANGELOG).

Updated the Release Guidelines to recommend using this script.

* Add script to update the index.js JSDoc

Add a simple script that can be invoked to update the JSDoc in index.js.
The script simply reads the version in the package manifest and writes
it to the JSDoc of index.js. It does not check whether this changes
anything or not.

Updated the Release Guidelines to recommend using this script.

* Update guidelines to bump version

Update the "Updating the Version Number" section of the Release
Guidelines to include a command that can be used to apply a version bump
in the manifest and lockfile.

Additionally add subheaders to the section for readability.

* Add job to Automated Release workflow to commit a version bump

Add a new job to the Automated Releases workflow to automate the version
bump commit. To prevent having to push to the `main` branch (allowing
for full branch protections) and to encourage review before accepting an
automated change, the version bump is committed to a separate branch and
a Pull Request is opened through peter-evans/create-pull-request [1]. It
uses the new scripts added in the Release Guidelines to make the changes
for the version bump commit.

The existing jobs in the workflow are updated to only run on `push`
events, not changing their behaviour otherwise. The new job is only run
on `workflow_dispatch`, preventing it from running when a release is
being published.

I have tested the changes to the workflow in a private repository.

* Update the Release Guidelines

Largely rewrite the Release Guidelines. Simplify the explanation of
automated releases and expand & structure the manual release guide.

--
1. https://github.com/peter-evans/create-pull-request/tree/main