Fuzz escape function
Update the fuzz target to test both the `quote()` and `escape()`
functions (instead of just `quote()`). Fuzzing the `escape()` function
is roughly identical to fuzzing the `quote()` function, except that the
provided buffer is stripped of whitespace before echoing to ensure all
characters in the string are part of the first argument.

Some refactoring happened in index.fuzz.cjs related to this:
- Define `WHITESPACE_REGEX` to avoid rewriting the regular expression to
  (really) capture all whitespace.
- Update the error message for unexpected output so it's different for
  the two tested functions.
- Define the `options` for `quote()` and `escape()` prior to the resp.
  fuzz function as the value is the same for both.