Update unix escaping based on fuzzing
Update escaping for unix platforms based on the existing fuzz corpus.
The unix escape logic now expects a boolean value indicating whether
interpolation is enabled and if so will escape more characters. This is
needed when the `escape()` API functions is used because the result may
be used in a context where interpolation is enabled. Hence, main assumes
interpolation is enabled unless specified otherwise.