Always escape "*" and "?" on unix with interpolation
The "*" and "?" characters are both wildcard characters in arguments on
most shells on unix. As opposed to Zsh (where these characters were
already being escaped), bash and dash only expand these wildcard
characters when they actually match something. If nothing is matched
they're printed literally.

Since we can't know the context in which Shescape is used, it should be
assumed they can always be expanded, so these characters will always be
escaped (this is valid because "\*" always evaluates to the literal
string "*" and likewise "\?" evaluates to the literal string "?").