Fuzz the `escape` function, add escaping for interpolation characters (#152)
* Fuzz `escape` function on unix

Update the fuzz target to test both the `quote()` and `escape()`
functions (instead of just `quote()`). Fuzzing the `escape()` function
is roughly identical to fuzzing the `quote()` function, except that the
provided buffer is stripped of whitespace before echoing to ensure all
characters in the string are part of the first argument.

* Refactor index.fuzz.cjs to reduce duplication and fix Windows args

- Define `WHITESPACE_REGEX` to avoid rewriting the regular expression to
  (really) capture all whitespace.
- Update the error message for unexpected output so it's different for
  the two tested functions.
- Define the `options` for `quote()` and `escape()` prior to the resp.
  fuzz function as the value is the same for both.
-  Fix `prepareArg` for Windows cmd.exe when quoted=false

* Update unix escaping based on fuzzing

Update escaping for unix platforms based on the existing fuzz corpus.
The unix escape logic now expects a boolean value indicating whether
interpolation is enabled and if so will escape more characters. This is
needed when the `escape()` API functions is used because the result may
be used in a context where interpolation is enabled. Hence, main assumes
interpolation is enabled unless specified otherwise.

* Escape ")" for unix with interpolation

This escape relates to the escaping of "(" in that shells (tested Zsh) on
unix with interpolation interpret parenthesis in a special way.

* Escape "<" and ">" for unix with interpolation

These characters are used for redirecting input/output streams so should
be escaped.

* Escape "#" at the start of args for unix with interpolation

The "#" character has a special meaning at the start of an argument.

* Escape "|" in args for unix with interpolation

The "|" character means a logical or.

* Escape ";" in args for unix with interpolation

The ";" character denotes the end of a singular command, after which a
new command starts.

* Escape "&" in args for unix with interpolation

The "&" character means a logical and.

* Escape "*" and "?" in args for unix with interpolation

The characters  "*" and "?" can be used for string expansion if
there are no quotes.

In Zsh these characters are always expanded (if unsuccessful the
command will error). On bash and dash these characters will only
be expanded if possible (otherwise they will appear literally). As
the context in which Shescape operates is unknown, for both shell
styles the characters always need to be escaped.

* Escape leading "~" in args for unix with interpolation

The "~" character has a special meaning (the home directory) either:
1. as the only character in the argument (bash, dash)
2. at the start of an argument (Zsh)
Both cases can be efficiently escaped by prefix "~" with a backslash,
this works regardless of whether or not the argument is more than one
string in bash and dash.

* Escape "[" and "]" for Zsh with interpolation

Specifically in Zsh, "[" and "]" can be used for string expansion if
there are no quotes. So, the err on the safe side, they will always be
escaped when `interpolation` is true.

* Escape leading "=" for Zsh with interpolation

The "=" character has a special meaning at the start of an argument in
Zsh.

* Escape "{" and "}" for Zsh with interpolation

Specifically in Zsh, "{" and "}" can be used for string expansion if
there are no quotes. So, the err on the safe side, they will always be
escaped when `interpolation` is true.

* Fuzz `escape` function on PowerShell

Run the fuzz target with the existing corpus on PowerShell and fix all
problems that are uncovered.

* Escape single quote variants in PowerShell on Windows

Escape the following characters for PowerShell on Windows because it
will interpret these as regular single quotes.

- U+2018 (Left Single Quotation Mark)
- U+2019 (Right Single Quotation Mark)
- U+201B (Single High-Reversed-9 Quotation Mark)
- U+201A (Single Low-9 Quotation Mark)

* Escape "<" and ">" for Windows PowerShell with interpolation

These characters have special meaning when they appear at the beginning
of an argument. In the case of ">", it also has this meaning when it is
prefixed by "1", "2", "3", "4", "5", "6", or "*".

* Escape "@" for Windows PowerShell with interpolation

This character has special meaning when it appears at the beginning of
an argument.

* Escape "]" for Windows PowerShell with interpolation

This character has a special meaning when it appears at the beginning of
an argument.

* Escape "," for Windows PowerShell with interpolation

This character is used to separate commands on PowerShell.

* Escape leading "-" in PowerShell on Windows

If "-" is the first character in an argument in PowerShell on Windows it
has a special meaning, so it's always escaped.

* Escape leading ":" in PowerShell on Windows

If ":" is the first character in an argument in PowerShell on Windows it
has a special meaning, so it's always escaped if it is the first
character.

* Fuzz `escape`  function on cmd.exe

Run the fuzz target with the existing corpus on cmd.exe and fix all
problems that are uncovered.

* Escape "^" for Windows cmd.exe with interpolation

This character is used for escaping, so it must be escaped itself. As it
is used for escaping, it should be escaped first to prevent escaping the
"^" instances inserted to escape other characters.

* Update unit test suites for changes to unix.js and win.js

* Update unit test suites for changes to main.js

* Update TypeScript type definitions

* Update documentation for `escape` and `escapeAll`

* Update CHANGELOG