Refactor unix.js and win.js to support many shells
For both modules, define a `Map` that maps the names of shell binaries
to a respective function that escapes arguments for that shell.

For win.js this amounts to replacing the if-else-statement in
`escapeShellArg` by logic that obtains the escape function from the map.

For unix.js more changes were required. In contrast to win.js the unix
module wasn't yet set up to separate argument escaping by shell. Hence,
first the individual escape functions are rewritten such that there is
one per (group of) shell(s). Second, the same map-related logic from
win.js is added to the `escapeShellArg` function of the unix module.

Both of these implementations use the standard NodeJS `path` package to
reliably obtain the name of the shell binary from the provided shell,
which may be a full path. As this is a new import, the Rollup config is
updated to mark it as an external dependency.