Resolve (sym)links before returning the shell
In `getShell`, instead of just returning the provided/default shell, try
resolve it as a (sym)link first. This is very useful on unix systems,
where the default shell is "/bin/sh", but "/bin/sh" is often a symbolic
link to the distro's default shell.

Note that, before resolving (sym)links the system will check if there is
a file for the binary. This is because the function call for resolving
(sym)links will throw an error both when the file doesn't exist and when
the file isn't a (sym)link. In the future (because we need to maintain
backwards compatibility now) the system may need to throw an error in
case the shell doesn't exist (because we want to fail closed to prevent
a scenario where the shell argument is not escaped correctly because the
shell couldn't be determined correctly).