Add dependency `which` to be able to expand shells to their full path
Add `which` [1] as a dependency. This dependency can be used to expand a
shell-as-a-string to its full path. This is necessary because we want to
check if the shell exists and, if it does, potentially resolve it as a
(sym)link.

Update the README to remove "Zero dependencies" as a feature.

--
1. https://www.npmjs.com/package/which