Refactor logic to resolve the location of the shell into a module
Refactor the logic to find the "true" executable of a shell out of
`./main.js` and into a new module. This module is written more
generically and is fully tested (verified with coverage and mutants).

The module uses explicit dependency injection to allow for testing it
thorougly without having to rely on the underlying filesystem and
environment variables. Outside tests, `./main.js` just provides the
required dependencies whenever it calls the `resolveExecutable`
function.