Focused test timeout extension for main.test.js smoke test
Provide extended timeout (from 2s to 5s) to the smoke tests for
`main.test.js`. These tests run for quite a few different combinations
of inputs, and coupled with having to make system calls, this can be
somewhat slow. We're seeing some timeouts on slower Windows machines so
are providing a longer timeout window for specficially the affected
tests. For generality all smoke tests (even those for unix systems,
which don't seem to be affected too badly) are included.