Follow symbolic links when determining the shell to escape for (#159)
* Add dependency `which` to be able to expand shells to their full path

Add `which` [1] as a dependency. This dependency can be used to expand a
shell-as-a-string to its full path. This is necessary because we want to
check if the shell exists and, if it does, potentially resolve it as a
(sym)link.

For maximum compatibility, Shescape supports any v2 release of `which`.

* Add Sinon.JS for testing with Stubs

Add Sinon.JS [2] as a devDependency for testing. Sinon.JS provides an
API for creating, managing, and verifying the behavior of spies/stubs/
mocks. This is used in the test suite for the new `::resolveExecutable`
function.

* Resolve (sym)links before using the shell

In `getShell`, instead of just returning the provided/default shell, try
resolve the shell (provided or default) as a (sym)link first. This is
very useful on unix systems, where the default shell is "/bin/sh", but
"/bin/sh" is often a symbolic link to the distro's default shell.

Before resolving (sym)links the system will first try to resolve the
shell string to its full path using `which` [1] and second check if the
file for the shell binary actually exists. Both of these actions fail
open - just returning the shell string in case of a failure - to provide
backwards compatability. The first check is needed because in order to
follow a (sym)link you need its full path. The second check is needed
because the `readlink` function fails both when the file doesn't exist
and when the file is not a (sym)link, hence by checking if the file
exists first the failure scenario is reduced to just "it's not a
(sym)link".

This logic is written in a new function in a new module to improve its
testability. The logic is quite complex, and relying on the real system
calls may result in flaky tests. For full control, dependency injection
is used for the function interface.

* Focused test timeout extension for main.test.js smoke test

Provide extended timeout (from 2s to 5s) to the smoke tests for
`main.test.js`. These tests run for quite a few different combinations
of inputs, and coupled with having to make system calls, this can be
somewhat slow. We're seeing some timeouts on slower Windows machines so
are providing a longer timeout window for specficially the affected
tests. For generality all smoke tests (even those for unix systems,
which don't seem to be affected too badly) are included.

* Update the README and the CHANGELOG

--
1. https://www.npmjs.com/package/which
2. https://sinonjs.org/