Reduce cognitive complexity of main.js
Reduce the cognitize complexity and number of functions in main.js
following the design of the `resolveExecutable` function. A new function
is added, `escapeShellArg` that is explicitly provided some functions
that can come from eiterh the unix.js or win.js module. By doing this,
most of the branches related to the platform are now gone (one is kept
because it's for a feature that is marked for removal in the next major
version).

To accomodate this, the unix.js and win.js now expose a `getBasename`
function. Additionally, unix.js and win.js were refactored to explose a
function called `getEscapeFunction` instead of the
`escapeFunctionsByShell` map directly, primarly to avoid accidentally
changing the map outside of the respective modules.

All existing tests were changed to accomodate these changes. The
`getBasename` functions are independently tested as well.