Support different quote functions for different shells
Refactor main.js, and unix.js and win.js accordingly, to be able to use
a different function for quoting an argument based on the selected shell
(rather than just the platform).

All changes are fully reflected in the test suites and all changes to
the behaviour of functions are tested completely.