Refactor codebase for better scalability (#161)
Refactor the codebase to make it easier to scale w.r.t. new platforms
and new shells. The main module now accepts a set of functions that,
depending on the platform, can be used to escape an argument. As such
both the main module and the per-platform modules (unix.js and win.js)
have been heavily refactored.