Improve win::getDefaultShell property tests (#198)
Add a new test to the property test for the windows module, specifically
for the `getDefaultShell` function that aims to test that, no matter the
environment variables, the default shell is *always* a string. This
property test is modelled after [1], which says that values on
`process.env` are always strings. Manual testing (on Node.js v10.13.0,
v12.0.0, and v16.13.1) confirmed this behaviour. Some examples of the
behaviour observed:

    process.env.foo = true;
    console.log(process.env.foo);
    // => 'true'

    process.env.foo = 12;
    console.log(process.env.foo);
    // => '12'

    process.env.foo = {};
    console.log(process.env.foo);
    // => '[object Object]'

    process.env.foo = undefined;
    console.log(process.env.foo);
    // => 'undefined'

    process.env.foo = null;
    console.log(process.env.foo);
    // => 'null'

    process.env.foo = NaN;
    console.log(process.env.foo);
    // => 'NaN'

The env arbitrary is constrained to model the observed behaviour. Plus,
it's modelled such that the %COMPSEC% environment variable is *likely*
to be present.

--
1. https://nodejs.org/api/process.html#processenv