Run CommonJS tests on all supported Node.js versions
And all "normal" tests only on the development Node.js version. This is
necessary because the source code is written in ESM, but Node v10.13.0
does not support ESM, so running the "normal" tests is non-trivial.
Running the CommonJS test suite on every supported Node.js version will
at least catch faulty imports and, to an extends, faulty implementations
for older Node.js versions.